requirements.txt
decimal>=1.70
python-dotenv>=1.0.0